Close fd after fork() also on REST API
Closing the file descriptors also on REST API after the fork.
It excludes all the pipes.
close_all_non_term_fd() accepts a vector of FDs to exclude
